+++ CHANGES.rst
@@ -4,6 +4,7 @@ IP Navigator CHANGES
 
 development
 ===========
+- Get rid of more static html templates and refactor to javascript application
 
 2017-10-09 0.138.0
 ==================